+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## UNRELEASED
 
+### FIXED
+
+- Various workflow errors
+
+## [sprint-08](https://github.com/amosproj/amos2025ss01-embark/releases/tag/sprint-08-release) - 2025-06-11
+
 ### ADDED
 
 - Update/Reset buttons for configurations ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=112359403&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C42))
@@ -9,6 +15,7 @@
 - Worker soft reset functionality ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=112364355&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C43))
 - Worker update functionality ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=112364538&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C45))
 - Celery setup ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=113497561&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C55))
+- Periodic worker information updates ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=112364821&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C47))
 
 ### FIXED